pcs trait impl file too large, separate into 2 files
